move "format_as_xml"
