+++ docs/graph.md
@@ -359,8 +359,7 @@ from dataclasses import dataclass, field
 
 from pydantic import BaseModel, EmailStr
 
-from pydantic_ai import Agent
-from pydantic_ai.format_as_xml import format_as_xml
+from pydantic_ai import Agent, format_as_xml
 from pydantic_ai.messages import ModelMessage
 from pydantic_graph import BaseNode, End, Graph, GraphRunContext
 
@@ -663,7 +662,7 @@ Instead of running the entire graph in a single process invocation, we run the g
     )
 
     from pydantic_ai import Agent
-    from pydantic_ai.format_as_xml import format_as_xml
+    from pydantic_ai import format_as_xml
     from pydantic_ai.messages import ModelMessage
 
     ask_agent = Agent('openai:gpt-4o', output_type=str, instrument=True)
