+++ pydantic_ai_slim/pydantic_ai/format_prompt.py
@@ -42,7 +42,7 @@ def format_as_xml(
 
     Example:
     ```python {title="format_as_xml_example.py" lint="skip"}
-    from pydantic_ai.format_as_xml import format_as_xml
+    from pydantic_ai import format_as_xml
 
     print(format_as_xml({'name': 'John', 'height': 6, 'weight': 200}, root_tag='user'))
     '''
