+++ tests/test_format_as_xml.py
@@ -6,7 +6,7 @@ import pytest
 from inline_snapshot import snapshot
 from pydantic import BaseModel
 
-from pydantic_ai.format_as_xml import format_as_xml
+from pydantic_ai import format_as_xml
 
 
 @dataclass
